requirements/base.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.2
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-jwt==1.19.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.1
